+++ packages/scripts/src/commands/build_storybook/execute.ts
@@ -2,7 +2,8 @@
  * Copyright (c) Microsoft. All rights reserved.
  * Licensed under the MIT license. See LICENSE file in the project.
  */
-import { runSequential, getStorybookConfigPath } from '../../utils'
+import { getStorybookConfigPath } from '../../utils'
+import { runSequential } from '../../jobrunner'
 
 export interface BuildStorybookCommandOptions {
 	verbose: boolean
@@ -18,10 +19,8 @@ export async function execute({
 		args.push('--quiet')
 	}
 
-	return runSequential([
-		{
-			exec: 'build-storybook',
-			args,
-		},
-	])
+	return runSequential({
+		exec: 'build-storybook',
+		args,
+	})
 }
