+++ packages/scripts/src/commands/commit-msg/execute.ts
@@ -3,11 +3,12 @@
  * Licensed under the MIT license. See LICENSE file in the project.
  */
 /* eslint-disable @typescript-eslint/no-non-null-assertion */
-import { run, getConfigPath } from '../../utils'
+import { getConfigPath } from '../../utils'
+import { runSequential } from '../../jobrunner'
 
 export async function execute(): Promise<number> {
 	const commitLintPath = await getConfigPath('commitlint.config.js')!
-	const { code } = await run({
+	const code = await runSequential({
 		exec: 'commitlint',
 		args: ['--config', commitLintPath, '-E', 'HUSKY_GIT_PARAMS'],
 	})
