+++ packages/scripts/src/commands/precommit/index.ts
@@ -2,7 +2,8 @@
  * Copyright (c) Microsoft. All rights reserved.
  * Licensed under the MIT license. See LICENSE file in the project.
  */
-import { getConfigPath, run } from '../../utils'
+import { getConfigPath } from '../../utils'
+import { runSequential } from '../../jobrunner'
 import { Command } from 'commander'
 
 export default function preCommit(program: Command): void {
@@ -11,7 +12,7 @@ export default function preCommit(program: Command): void {
 		.description('execute pre-commit tasks (for husky hook)')
 		.action(async () => {
 			const configPath = await getConfigPath('.lintstagedrc')
-			const { code } = await run({
+			const code = await runSequential({
 				exec: 'lint-staged',
 				args: ['-c', configPath],
 			})
